ENH: Improve the itkScalarToRGBColormapImageFilter coverage
This patch set exercises the basic object methods of the filter. Since
these call the PrintSelf method, the explicit call to the Print method has
been removed.

The Jet and HSV colormaps are now set through the void
SetColormap(ColormapEnumType) function, and no longer through the
itkSetObjectMacro(Colormap, ColormapType) macro, as done with the rest of
the colormaps in the test. This will enable exercising the
itkScalarToRGBColormapImageFilter.hxx:238 and 246 to be called.

Change-Id: I6f17cb23c01e25c95b3d8225f397fa5c02d3fc1a