BUG: try_compile needs to have CMAKE_CXX_STANDARD passed
When using cmake version 3 or greater, and -DCMAKE_CXX_STANDARD:STRING=11
then the try_compile needs to explicitly have CMAKE_CXX_STANDARD passed
as an option so that the same compilation environment is used consistently
for introspection testing and the final product.

This patch is represented in the upstream VXL and should merge without
conflicts when VXL is upgraded in the future.

Without this patch set, VXL is always resulting with instrospection result
that claims CXX98 rather than CXX11 and then compilation errors result
due to mismatch of VXL_CONSTEXPR.

Change-Id: I24ee7d7e88abe994e0a19d2423fa32b81f12dcc0