COMP: CID-1266557 StatisticsLabelMapFilter uninitialized scalar field
This addresses Coverity issue CID 1266557: Uninitialized scalar field
(UNINIT_CTOR):

uninit_member: Non-static class member m_Minimum is not initialized in this
constructor nor in any functions that it calls.

uninit_member: Non-static class member m_Maximum is not initialized in this
constructor nor in any functions that it calls.

Change-Id: Ib42df517ffdc3a1b58229ddd62755eb4ce80d988