ENH: Move PrincipalComponentsAnalysis to itk-module hierarchy
This is the first part of a series of commit to convert this repository to
work with ITK's external modules infrastructure.
This commit allows to configure/generate the module against an ITK build.
The itkVectorFieldPCA.* files were moved to the folder named include which
the external module expects for header files.
The itk-module.cmake was added to describe the module and its dependencies.