ENH: Move PrincipalComponentsAnalysis to itk-module testing hierarchy
The external module needs its testing to be under the test folder. The
itkVectorKernelPCATest was moved accordingly.
This commit does not build. But since we moved a file from one folder to
another, we will make the appropriate change in the next commit so they
stand out as actual change inside the file.