ENH: New remote module GenericLabelInterpolator
This remote module integrates the method and code presented in [1] written by
Schaerer J., Roche F., Belaroussi B..
The basic idea of this generic interpolator for label images is to interpolate each
label with an ordinary image interpolator, and return the label with the highest
value. This is the idea used by the itk::LabelImageGaussianInterpolateImageFunction
interpolator. Unfortunately, this class is currently limited to Gaussian interpolation.
Using generic programming, our proposed interpolator extends this idea to any image
interpolator. Combined with linear interpolation, this results in similar or better
accuracy and much improved computation speeds on a test image.

[1] http://www.insight-journal.org/browse/publication/950

Change-Id: I9ee7372a54b7356a47fce9e95dce0b2bc242886b