BUG: ImportImageFilter not responsible of memory management
ImportImageFilter had an option to take ownership of the imported data. If
this option was set, the filter was releasing memory once it was
destroyed. This was creating problems if the output of the filter was used
in a different scope. To avoid this problem, the memory is managed
directly by the ImportImageContainer that is created when importing the
external data. This way, the memory management is independent from
ImportImageFilter, and the memory is only released if no object uses the
the output anymore.

Change-Id: I20270637b35ffad296b54e5ff89d51940915842e