BUG: Address itkHDF5ImageIO test failures for mingw64
The Windows compilers VS and MINGW64 still use unsigned __int64 for
itk::SizeValueType and not the unsinged long type. This patch addresses
reading and writing this type for HDF5 meta-data. Specifically the
dimension variable is an array of this type and the HDF5ImageIO was
failing to read or write this type with an exception.

ITK has assumed that the compiler supports the long long type since
ITK version 4.0. Therefore, this code block can safely be enabled for
all compilers ITK supports.

Change-Id: I0cec69e176787b043d5010e32a2658bf88f1c0ad