STYLE: Provide a well documented macro for repeated code
A pervasive paradigm in ITK for explicitly dis-allowing
copy constructors and operator= can be documented more
clearly in one spot of the itkMacro.h file and then
used consistently in the toolkit.

CMake facilities for providing fine grained compiler feature
control used:
ITK_DELETED_FUNCTION instead of
ITK_DELETE_FUNCTION  (Note the tense of DELETE)

This change was implemented to be consistent with other
packages that conform to the Cmake 3.5.2 compiler feature
paradigm designations.

Change-Id: Id43f4d78223effb0ab61b213e2c1b02689aa40a9