ENH: Reorganize examples and tests
When compiled as a remote module outside of ITK and inside the
ITK source tree, the behavior of the module should be the same.
Since the example requires the ITK IO factory mechanism, and to
follow the ITK guide lines, the 'examples' folder is now
a different CMake project: it is not included in the remote
module compilation.