BUG: ResampleImageFilter honor isInside output of the transform method
For the SpecialCoordinatesImage classes honor the isInside value returned
by TransformPhysicalPointToContinuousIndex.

For itk::Image classes, this value is not checked because the interpolator
IsInsideBuffer() method is used instead. IsInsideBuffer takes into account the
extra have pixel domain in physical space from the center of the pixel. the
itk::Image::TransformPhysicalPointToContinuousIndex does not take this into
account (and we would not want to add it for performance and backwards
compatibility reasons).

Due to the different behavior fo itk::SpecialCoordinatesImage, this extra
check needs to also be added.

Change-Id: I31f9556427c8ad22ce7033ac230ebd4ead09d6a8