BUG: Labled Point Set Registration: test multiple labels
The line modified was clearly intending to create multiple lables,
but due to the division factor chosen was labeling every point 1.
With this modification, points are split into 2 groups of unequal
size, labled 1 and 2. This is important to break the rotational
symmetry of the pointset.

Since the test uses an affine transformation, with all points labled
the same any rotation applied to the moving transformation is also
a valid registration. This can confuse the test since it depends
on a unique valid registration.

Change-Id: I4831d105316edd57dd5f1124d5db1bd4e5080354