COMP: VCL_DEFINE_SPECIALIZATION not defined
A build failure was encountered when building ITK against the
development branch of VXL (using `ITK_USE_SYSTEM_VXL=ON`).
The failure is due to ITK's use of a macro, `VCL_DEFINE_SPECIALIZATION`,
which has been removed from VXL in a recent commit.  Replacing
this macro with `template<>` avoids the build failure.

Change-Id: Ib20be95ed9b7d99ca797b2dcba4834fe422e55c2