ENH: Improve itkFileListVideoStyle.
Provide the missing ConstPointer standard class typedef.

Set the name of the superclass as the name for the itkTypeMacro RTTI
instead of the Superclass typedef.

Finish the implementation of the PrintSelf method: print the contents of
the m_FileNames ivar.

Remove implementation file out-of-method-body comment blocks,

Provide GetImageIO() and GetFileNames() methods.

Make PrintSelf be the last method of the class.

Improve the documentation style.

Fix some typos.

Change-Id: Ie01d8d2db88b349c56c90e351b27224165f9bb18