BUG: Confusing interface for FFTW wisdom files
The behavior of setting environmental variables and
using the available Functions was different.

If a user set the environmental variable
ITK_FFTW_WISDOM_CACHE_BASE=/tmp

==write phase==
Then the wisdom files are written to /tmp/<wisdom_file>

==read phase==
The wisdom files are also read from /tmp/<wisdom_file> during
the construction of the FFTWGlobalConfiguration singleton.

-*-*-*-*-*-*-*-*
If, however, the developer uses the API:
  itk::FFTWGlobalConfiguration::SetWisdomCacheBase("/tmp");

==write phase==
Then the wisdom files are written to /tmp/<wisdom_file>  [OK]

==read phase==
The wisdom files are initially read from the hard coded default of
searching in the users home account ${HOME}/<wisdomfile> during the
construction of the FFTWGlobalConfiguration singleton, BUT THE NEW
DEFAULT WISDOM PATH IS NEVER USED FOR READING AFTER calling
'SetWisdomCacheBase("/tmp");'

This patch accumulates the wisdom of all files found after each
setting of SetWisdomCacheBase() or SetReadWisdomCache(true), so that
the desired behavior of using the written wisdom fles is achieved.

Change-Id: I78e6b8e4cc23f9909b4628b44c6f150600e3b653