ENH: Add support of images types with the 'CovariantVector' pixel type
The 'GradientVectorFlowImageFilter' should support images types with 
the 'CovariantVector' pixel type. This is necessary for compatibility 
with other filters, such as the 'GradientImageFilter'.

Also we can see a note from ITK documentation:
"Note that itk::Vector is not the appropriate class for representing 
normals to surfaces and gradients of functions. This is due to the way 
vectors behave under affine transforms. ITK has a specific class for 
representing normals and function gradients. This is the itk::CovariantVector class."

Change-Id: Ib78847ae2d6d261f685bbea828e5d059399e6e9f