BUG: Match Add2 function to AddImageFilter documentation
Remove the undocumented cast of the first operand to the
AccumalateType. The implementation of the Add2 functor matched
neither the documentation for the AddImageFilter nor the ConceptCheck
performed to see if the pixel types support the operator.

This patch addresses the following error: itkAddImageFilter.h:54:40:
error: no match for ‘operator+’ (operand types are ‘const
AccumulatorType {aka const double}’ and ‘const std::complex<float>’)
     return static_cast< TOutput >( sum + B );

Change-Id: I8bb451c948b49970531f9bf9067c2bde8a50b817