ENH: Add ArchiveTestingDataOnGirder.py
This script is used to archive testing data on the data.kitware.com
girder instance. It requires fixes in Girder master. When passed an
ExternalData object store populated with SHA512 entries, it populates a folder
tree corresponding to the ITK folder hierarchy and *.sha512 content links.
It is intended to be executed on releases to archive the testing data on
data.kitware.com.

Change-Id: Idf50ef30b3912d265f71d322309786e7d8c1bf7f