BUG: Path to commit was relative instead of absolute
When printing the message that explains how to proceed to
create a patch to improve a third party library, the path
in the command line to create the patch was relative. It was
working if the user's current working directory was the top
level directory of the git project, but was not working
otherwise. To avoid this issue, the command line printed
to help the user now shows the absolute path. The command
can be run from any directory from within the git project.

Change-Id: I30ab990335edc0d7eb310ebc738aa3a6b8f14ea2