ENH: Path to SetupForDevelopment is relative, not absolute
When pre-commit hook reports an error because the version of
SetupForDevelopment.sh with which the project was configured
is too old, the script prints an error message telling the
developer what command to run to update their configuration.
However, the command showed to the developer shows a hard-coded
relative path that may not correspond exactly to the command
that the developer must run. Instead, this patch replaces
the relative path by an absolute path. The printed command line
will always be accurate.

Change-Id: I58664dff8f27a5a93e38f1f3f7d9813ea62b85db