BUG: Missing associated test data
When running tests requiring the file 'mri3D.mhd',
this file was added as a file with no associated data file.
However, mri3D.mhd does not contain data, but is merely
a header file. The image data is contained in:
DicomSeries/Image%04d.dcm 75 77 1
The data file and its associated files are in different
directories, respectively ${ITK_DATA_ROOT}/Input/ and
${ITK_DATA_ROOT}/Input/DicomSeries/
Since DATA{} does not allow to specify an associated
file that is in a different directory, we call ExternalData_Expand_Arguments
manually to download the required files.

The tests using mri3D.mhd currently seem to be working but
they only pass because other tests download the
data files before these specific tests are run.
If ITK is built with only a few
modules ON, such as ITKIOMeta and ITKMetaIO,
the tests requiring mri3D.mhd will fail.

Change-Id: I86576d30bd1b50bb4bc438a2ab88842a4c6e8d1b