BUG: Quotes are not escaped in configured cmake file
When configuring a module that uses VTK, the VTK macro
vtk_module_config is called. This macro configures several variables,
including ${namespace}_DEFINITIONS. If the project is configured
for Visual Studio, this variable contains paths which contain
quotes. When using this variable to configure the module, the
configured file was not escaping these quotes, thus creating
warning messages which indicate an error in this case:
e.g: set_property(DIRECTORY APPEND PROPERTY COMPILE_DEFINITIONS
" vtkRenderingCore_INCLUDE="D:/workspace_itk_org/InsightToolkit-4.10.1-bin/\
CMakeFiles/vtkRenderingCore_AUTOINIT_vtkInteractionStyle_vtkRenderingFreeType\
_vtkRenderingOpenGL2.h" ")

The quotes inside the quotes create problems.
To avoid this issue reported here [1], the variable is not expanded
in the configured file, but kept as a variable. It is expanded only
when required.
This follows the same pattern as vtkGlue [2]

[1] http://public.kitware.com/pipermail/community/2016-October/012141.html
[2] https://github.com/InsightSoftwareConsortium/ITK/blob/master/Modules/Bridge/VtkGlue/CMakeLists.txt#L100

Change-Id: I9edc4a263a04bdca29dc5dbcb7fed6fb4704db78