COMP: Fix changed Graft method
Move the helper function for LabelGeometryImageFilter to a private
namespace. These are not public methods.

Remove casts in CalculateOrientedImage, by using an TInputImage
type (formally TGeneric ) and output image type directly as template
parameters to the ResampleImageFilter. This remove use of Graft for
types that were not used by an "if" statement.

Co-Author: Francois Budin <francois.budin@kitware.com>

Change-Id: I15df0fff477641b9cc227d81e8c138a4f2385e78