BUG: Swig and PCRE do not support space in installation paths
PCRE and Swig both throw errors when trying to compile ITK in
a directory that contains some spaces. The error can maybe be removed
for PCRE if the project is configured with CMake (it contains a CMakeLists.txt
file) instead of configuring it with "configure". Correcting that problem
for Swig seems more difficult. Instead of having an error at
compilation time, it is cleaner to warn the user about the problem
and not let them continue.
The error for Swig is documented for Windows [1] but is reproducible
on Linux.

[1] http://www.swig.org/Doc3.0/Windows.html

Change-Id: Ice3f931cf701dedbb98ce1f4291d8a31e8f847f2