BUG: Use BSpline Order 3 for 2D in transform factory.
This BSplineTransform's 3rd parameter is the spline order and NOT the
output dimension order. The default value is 3, and other spline
orders are rarely used.

This spline order template parameter is neither encoded in the Fixed
parameters or in the Name of the class.

This patch changes the BSplineTransform created by the
TransformFactory from order 2 to order 3 for 2d. Since it is
impossible to know the intended order of the written BSpline, it is
presumed to be 3 and that is the default and standard order to use.

Change-Id: Ic32de1be8caca0294c6aef801c0dc0593db9e998