BUG: Use NATIVE_HBOOL for encoding HDF5 types.
As reported in ITK-3484,

bool, long and unsigned long are currently encoded using NATIVE_HBOOL which
happens to be equivalent to NATIVE_INT in HDF5 1.8.x.
In HDF5 1.10, NATIVE_HBOOL is equivalent to NATIVE_UCHAR and isn't suitable
for long or unsigned long anymore.

This fixes itkHDF5ImageIOTest test failure when building against HDF5-1.10.

Change-Id: I07fe2be64fc516e3b294f7ad4418d4e5ce69c9a8