STYLE: Improve the itkAntiAliasBinaryImageFilter style.
Make the constructor be the first method of the implementation file.

Make PrintSelf be the last method of the implementation file.

Use initialization lists.

SmartPointers default themselves to the null pointer.

Change the case of the test name to make it coherent with the naming
convention of the rest of the ITK tests.

Change-Id: I9cafd9471354f29f1fbb6e15b653354ddabbfbb0