BUG: Fix uninitialized variable due to dynamic analysis defect.
The m_InitialTransformParameters and m_LastTransformParameters ivars in
the itk::PointSetToPointSetRegistrationMethod class constructor were not
being initialized.

They were supposedly initialized using initialization lists, but for a
ParameterType type that does not indeed initialize the parameter; it makes
its size to have that value. A later call in the constructor made the
above variables be resized to a single component, but again, the
parameters were not being initialized.

Change-Id: Idec66e5cd316d407f1a9fad0fc7c1bdded17fe07