ENH: Improve the itkPointSetToPointSetRegistrationTest.
Remove a call to SetInitialTransformParameters using an identity transform
because the transform parameters are set at a later stage to some other
supposedly more appropriate values for the test.

Make the loops to set the transform parameters use the dimensionality of
the point sets (i.e. do not use a hard-coded limit condition in the loops).

Change-Id: Iac998da0f886db171ca338b0c30c440a5c0a5a16