BUG: Fix uninitialized variable defect reported by valgrind.
The m_InitialTransformParameters and m_LastTransformParameters ivars in
the itk::PointSetToImageRegistrationMethod class constructor were not
being initialized.

They were supposedly initialized using initialization lists, but for a
ParameterType type that does not indeed initialized the paremeters; it
makes its size have that value. A later call in the constructor made the
above variables be reiszed to a single component, but again, the
parameters were not being initialized.

Change-Id: Ia67245dbb1740d55e28996966df58748d3c0dcc5