ENH: Improve the itkPointSetToImageRegistrationTest.
Remove a call to SetInitialTransformParameters using an identity transform
because the transform parameters are set at a later stage in a loop
(although they are effectively set to the identify translation transform).

Make the loop to set the transform parameters be limited to the parameters
variable size, and not the image dimensionality, although they should
possibly be the same.

Change-Id: I992e2185f8ae5b98377307cd093f91543442e8ee