ENH: Provide mechanism for enhanced NIFTI support
NIFTI has several mechansims for describing the voxel
index to physical space orientation.  NIFTI also has
several ways to specify the intent of those transforms.

Display tools may interpret the different intent codes
and provide different processing capabilities dependant
on those settings.  Within ITK environment this was always
consistently managaged.

This patch set provides a mechanism that uses the meta-data
in the MetaDataDictionary to allow more control over the
meta-data that is stored in the NIFTI file.

Test cases were improved to ensure that the desired behavior
is achieved.

A tool developer may wish to ONLY use the qform (or sfrom) variant
 of codes in order to avoid ambiguity in how external tools
should interpret the index to physical space transforms.  This is now
possible by adding the appropriate information to the images
MetaDataDictionary before writing.

For example
  //Set the qform and sfrom codes for the MetaDataDictionary.
  itk::MetaDataDictionary & thisDic = img->GetMetaDataDictionary();
  itk::EncapsulateMetaData< std::string >( thisDic, "qform_code_name", "NIFTI_XFORM_SCANNER_ANAT" );
  itk::EncapsulateMetaData< std::string >( thisDic, "sform_code_name", "NIFTI_XFORM_UNKNOWN" );

Change-Id: I5689f23e0b55e39cf0a1a8c7b221f3d7571c0831