BUG: Use NIFTI sform/qform conventions correctly
The previous configuration for NIFTI files is OK if we only live in an
ITK echo-system.  In the ITK echo-system we ignore both the sform &
qform the intent codes because we only use the values in the qform.
These two lines have intent codes that are EXACTLY BACKWARDS from the
recommendations

https://github.com/InsightSoftwareConsortium/ITK/blob/master/Modules/IO/NIFTI/src/itkNiftiImageIO.cxx#L1717
https://github.com/InsightSoftwareConsortium/ITK/blob/master/Modules/IO/NIFTI/src/itkNiftiImageIO.cxx#L1718

The recommendations are provided:

https://nifti.nimh.nih.gov/nifti-1/documentation/nifti1fields/nifti1fields_pages/qsform_brief_usage
The qform_code should be set to either
   NIFTI_XFORM_UNKNOWN or
   NIFTI_XFORM_SCANNER_ANAT.
The sform code should be set to either
   NIFTI_XFORM_UNKNOWN,
   NIFTI_XFORM_ALIGNED_ANAT,
   NIFTI_XFORM_TALAIRACH or
   NIFTI_XFORM_MNI_152.

This patch swaps the default values for the sform and qform codes.

Change-Id: I0fd9a89ef14cb14a6f3ea19d432e7e88d89f8719