ENH: Remove ambiguity about sform/qform
ITK physical space representation matches the DICOM
standard of providing a rigid registration that is
representative of the anatomy in a scanner centric
space.  This most closely matches the qform intended
intent code of NIFTI_XFORM_SCANNER_ANAT.

To remove ambiguity of the intended intepretation of
a nifti file written by ITK, only the qform should
be used.  The sform adds no additional information
to understanding pysical space for images read/written
by ITK.

The sform is not required to be orthogonal, but ITK
only works with orthogonal physical space representations.

Change-Id: I62b43f49419a6b559bf9f43a2df081fe8b1bec1f