BUG: Ensure CreateImageIO is concurrent thread safe
Since not all ImageIO's are concurrent thread safe, and the
CreateImageIO may initialize all ImageIO, when the method checks
ImageIO::CanRead/Write method it may not be concurrent thread
safe. The conservative option to only allow one thread in the code
block to check all ImageIO's is done.

This solution is related to the following issue:
https://github.com/SimpleITK/SimpleITK/issues/74

Change-Id: I22aad01ce4208778c536d1588f690f7e4bb32301