ENH: Add CMake variable for max transform dimension, defaults 4
Add a new advanced CMake variable which specifies the maximum
transform dimension registered with the TransformFactory for IO.

The TransformFactory library is the largest in ITK because it
instantiates a large number of transform and related classes. The
default dimension is reduced from 9 to 4, reducing the size of the
library by up to 3 megabytes.

Move TransformFactoryTest to correct module. Update test to only check
for registered transforms if the are less that the max transform
dimension.

Change-Id: I7649afafbfb26eb9d4a104988b3a23d482cf072f