BUG: Supported extensions by PNG IO were not added in constructor.
Supported extensions by IO should be added in the constructor and
not in the CanReadFile() method. Adding them in the CanReadFile()
requires to call this function first to initialize the list of
supported extensions. Additionally, the list of supported extension
would be added to the current list every time CanReadFile() is called.

Change-Id: I28083b26a7005410dca860b229e38bc1cbed8b3a