ENH: Improving setting of default ITK_TEMPATE_VISIBILITY_DEFAULT
Connecting the initial value of ITK_TEMPLATE_VISIBILITY_DEFAULT to the
results of the compile and run of the the itkCheckPrivateDynamicCast
CMake module. The comment for this variable is also shortened and now
requires to the FAQ, for more information.

Allow the user to override ITK_TEMPLATE_EXPORT, by defining it before
include itkMacro.h.