BUG: Respect Visibility Preset for initial template visibility
If the user has explicitly set CMAKE_CXX_VISIBILITY to "hidden" and
static libraries are being used, then the intention is to hide all ITK
symbols and not have them exposed in the user's application or
libraries. This patch sets the default value of
ITK_TEMPLATE_VISIBILITY_DEFAULT to OFF for this case to respect the
users intent.

Change-Id: If7095ae5d54efebf08ca7eb67fa1395005b9ffe0