BUG: Enable program global timestamp
As demonstrated by the test in Change-Id:
I8deb652f2e117240a2092f96ae376920fd4b57da, the GlobalTimeStamp is not a true
singleton across the program when ITK is built with static libraries and
multiple binaries link to libITKCommon.

A current known configuration that this effects is the Python wrapping with
static ITK libraries. Each CPython extension module will get its own copy of
the GlobalTimeStamp used in itk::TimeStamp::Modified().

This patch addresses the issue by making the GlobalTimeStamp a pointer to a
static member of the itk::TimeStamp class. This pointer can be set and
obtained via accessor methods. When ITK is built with static libraries, the
Python module initialization code sets the GlobalTimeStamp pointer to the
pointer used in the _ITKCommonPython module.

A GlobalTimeStampInitializer (inspired by the singleton pattern in CTK's ctkSingleton.h)
is added to itkTimeStamp.cxx. The purpose of this class is to simply clean up
(delete) the binary's m_GlobalTimeStamp on program completion when a static
GlobalTimeStampInitializer's destructor is called.

To ensure that the global time stamp is created when needed without
performance penalty,

  static GlobalTimeStampType * globalTimeStamp  = GetGlobalTimeStamp();

is called in itk::TimeStamp::Modified(). This ensures that a time stamp is
created on demand (which could come from IO factory initialization code), but
that this method is only called once. Performance benchmarking with the
ITKPerformanceBenchmarking example benchmarks have shown no significant
difference with Linux / GCC 5.2 / Release CMAKE_BUILD_TYPE.

At the end of itkTimeStamp.cxx,

  static TimeStamp::GlobalTimeStampType * initializedGlobalTimestamp =
  itk::TimeStamp::GetGlobalTimeStamp();

is called to ensure that even if IO factory initialization code or other
initialization code does not initialize m_GlobalTimeStamp, m_GlobalTimeStamp
has been initialized once the library is loaded.

With static libraries, _ITKCommonPython module creates a C API to expose its
itk::TimeStamp::GetGlobalTimeStamp() method. Modules that link to libITKCommon
call their itk::TimeStamp::SetGlobalTimeStamp method with the return value of
this function. The PyCapsule mechanism, recommended in the CPython
documentation, is applied.

Change-Id: Ie1d8d2d66c99369b33cb782232bdc0c9bfa6d9a2