ENH: FindPythonLibs does not support both debug and release python libraries
This bug has been fixed in CMake > 3.7.1 [1].

Finding Python library paths using FindPythonLibs provided by CMake
results in an error if FindPythonLibs is included multiple times and
if multiple configurations of the Python library are available on the
system.

In ITK, this can happen if ITK_WRAP_PYTHON and Module_BridgeNumPy are
both selected.

PYTHON_LIBRARY may contain a list because of SelectLibraryConfigurations. If it
is the case, the list can be:

  optimized;<FILEPATH_TO_RELEASE_LIBRARY>;debug;<FILEPATH_TO_DEBUG_LIBRARY>

In this case, we need to set PYTHON_LIBRARY to one value, otherwise
get_filename_component will crash (too many arguments).
If SelectLibraryConfigurations was run, then the individual
PYTHON_LIBRARY_<CONFIG> values have been populated.

This patch sets the value of PYTHON_LIBRARY to only one value if necessary.
This allows to configure ITK with both ITK_WRAP_PYTHON and
Module_BridgeNumPy ON on Windows with both the release and the debug
version of the Python library installed.

[1] https://github.com/Kitware/CMake/commit/58dc3745ec925923713f16c0e77bc9277b29eb7d
Change-Id: Id512a0840613dbe154ccda902b12b212771a19ca