BUG: Use IsLinear to check if transform is "affine"
The prior implementation did not classify composite transform of
affine transform as "affine". By using the virtual method IsLinear,
the existing checked transforms, and the appropriate compositions of
affine transforms will be classified as "affine"

For the estimation of scale parameters for "affine" transforms the
corner sampling strategy is sufficient to estimate the scales. With
the recent change to the seeding of the random number generator to the
wall clock, there has been an randomness added to each run of the
scales esitmator. These estimated scales have a significant impact on
the optimization of the registration. By fixing this bug to use the
deterministic corner sampling the reliability and deterministic
behavior is improved.

Update baseline with for changes resulting from new optimizer scales
algorithm and seeds.

Change-Id: If37141fd349b4deece63386b5d4042761451ef6d