ENH: Do not use global random generator
The MersenneTwisterRandomVariateGenerator class is not concurrent
thread safe. Yet it was being used with a "Singleton-style" pattern
where all random iterators were using the same global
instance. Iterators assume that each instance is locally
thread-safe, that is separate instances can be used concurrently.

This results is the random iterators not being locally thread-safe,
and global side effects from using them. That is after setting a
"global" seed the results would depend on ordered usage of these the
global generator. Additionally, these iterators have "SetSeed"
methods, which are assumed to only have local ( just the iterator )
side-effects, and not modify the global generator's seed.

This changes the iterators to use a local instance of the random
generator, seeded with a deterministic permutation from the global
seed.

Change-Id: I62491d02bf98d41b86a69cf15f2f11a18a96520a