ENH: Update BridgeNumPy to latest version on 2017.01.08 to address ITK-3511
This update addresses bug report ITK-3511 [1]. BridgeNumPy cannot create an
ITK image directly from a transpose of a NumPy array. The array is not C-style
contiguous when computing the transpose. Memory cannot be easily shared between
Python and C++. A specific error message letting the user know what the problem
is is now displayed.

Francois Budin (2):
      ENH: Print useful error message if NumPy array is not C-style contiguous
      Merge pull request #12 from \
          fbudin69500/ENH_better_error_message_for_fortran_data_ordered_arrays

[1] https://issues.itk.org/jira/browse/ITK-3511

Change-Id: I82fcd4558e8169844ff5ae796c47ae4e5251ade2