BUG: Add option for fixed seed or wall clock seed for sampling
Adding two methods to set the seed for certain sampling strategies
used in the ImageRegistrationMethodv4. This allows setting a fixed seed
for deterministic behavior, so that multiple executions can have the
same point set. Alternatively, the pseudo-randomness can be seeded
with the wall clock at the start of each usage.

The default behavior is to seed from the "global" seed sequence, which
can be set to a fixed seed to enable deterministic behavior. With the
new GetNextSeed method, each ImageRegistrationMethodv4 object will
be initialized with a different seed.

A registration baseline with only 24 samples with translation
transform registration was updated.

Change-Id: I516d0c07ea4e7975c235f6ce4f30df622fdcda34