ENH: Expose GetNextSeed method in MersenneTwiseter Generator
Refactor constructor to expose a method which generates the "global"
seed sequence. This method is now used for construction of new
Generator objects. Also the v3 ImageToImage metric is to use this
result as opposed to always the global seed. This enables each new
metric to have distinct pseudo-randomness ( yet deterministic from the
global seed), which is consistent with prior behavior.

Change-Id: I4c0f3312efb7bbacbd0adebc79083c759e05c0ad