BUG: Binary include directory added only if enabled shared
If module is built with ENABLE_SHARED set to ON, an export header is generated
and written in the include folder in the binary directory. If module is not
built with ENABLE_SHARED ON, this include directory does not exist. When using
this module, an error will be generated because this include directory cannot
be found.

Change-Id: I53d7a2e6c64571936ec3b799e5bba96b8915ac3b