BUG: First IterationEvent is at CurrentIteration==0
All ITKv4 optimizer invoke the first iteration event when the
CurrentIteration is 0. This enables a hook at this iteration for
callbacks. This patch uses the CurrentIteration ivar as the loop
control variable for the optimization to provide consistent behavior
of when the iteration even is invoked, and removes the unneeded iter
variable.

Change-Id: I48b82d79f1e97247ca752706fa52c961de08bc3e