BUG: Address uninitialized variable defects.
Address uninitialized memory defects detected by Valgrind.

The m_SplinePoles variable was being created with a fixed length of 3.
However, depending on the spline order, only one or two of its elements
were being initialized at most.

This patch set changes the variable from a 3-element double array to an
std::vector of doubles, whose size is determined depending on the value of
the spline order. Once the size is determined, the corresponding elements
are initialized.

Change-Id: I42f883f08fccf0605251c1f93d2ef2cc23cb67da