BUG: BSplineScatteredDataPointSetToImageFilter output image is not generated
itkBSplineScatteredDataPointSetToImageFilterTest2 and
itkBSplineScatteredDataPointSetToImageFilterTest5 do not generate any
output image. On most computers, the output image will be initialized
to an image filled with 0, but that should not be expected. Since the output
image is not generated, we do not expect anything from the output image
and therefore do not compare it to a baseline.

Image size in itkBSplineScatteredDataPointSetToImageFilterTest3 was not
always correct because of rounding errors.

Change-Id: I27c444bd0bb628c20b5b03ae112037cfd819848b