BUG: Usage of std::atomic was incorrectly used in itk::Detail::AtomicOps
The class itk::Detail::AtomicOps is an implementation of atomic operations.
A previous patch [1] was aiming at using the standard atomic implementation
when it is available instead of using the ITK implementation.
However, the previous patch had introduced the usage of std::atomic
inside itk::Detail::AtomicOps. It was necessary to internally convert
back and forth between itk::Detail::AtomicOps and std::atomic.
This patch replaces itk::Detail::AtomicOps by std::atomic if the
standard class is available. ITK API is unchanged and atomic operations
are still exposed through itk::AtomicInt.

The previous implementation lead to the following issue: compiling ITK
with -std=c++14 flag was working, but the following error would arise when
compiling ITK without -std=c++14 flag and a project that depends on ITK that
was compiled with -std=c++14:

  undefined reference to `itk::Detail::AtomicOps<4ul>::Load(\
    itk::Detail::AtomicOps<4ul>::AtomicType const*)'

Additionally, this patch updates std::atomic support in the compiler info files
based on [2].

[1] Commit 0a64048aa8edc641428e9e43e0e57e8e3cd18fe4
[2] http://en.cppreference.com/w/cpp/compiler_support

Change-Id: Ibcce1d878f22e332b8f968efdf5843b38ab69bb1