DOC: Make it more explicit that 'push' authorization is not for everybody
When using the script 'SetupForDevelopment.sh', the user is asked if they
want to test push access. Despite the note written above this prompt,
some users are confused and concerned about the error message that is
printed if push access fails. This commit aims at clarifying that
typical users should not test push access.

Change-Id: I4f54f3428160a3b8f8d2049c6e90f8767624febe