ENH: Move morphological watersheds out of ITKReview
These are mature classes whose watershed segmentation implementation is
improved over the classes in the ITKWatersheds module.

The work was originally contributed in the Insight Journal paper:
"The watershed transform in ITK - discussion and new developments"
by Beare R., Lehmann G.
https://hdl.handle.net/1926/202
http://www.insight-journal.org/browse/publication/92

The supporting morphological classes are added to the
ITKMathematicalMorphology module and the segmentation classes are added to an
ITKWatersheds module.

Change-Id: I16d3b4af59789233c5c004d9a6e4fef9127033ef