BUG: 3519 KernelTransform WMatrix initialization
Kernel transform derived class need to initialize the matrices used
to compute the transform. When loading such a transform from a file,
the initialization was not done, which lead to crashes if the
developer was not manually calling the method ComputeWMatrix().
This patch aims at fixing this issue and calling the necessary
initialization method.

[1] https://issues.itk.org/jira/browse/ITK-3519

Change-Id: I07bd86bd72c873b313031fd78479847bf8f59b43