ENH: Improve the Watershed module code coverage.
Exercise basic object methods.

Exercise all Set/Get methods using the TEST_SET_GET_VALUE macro. Remove
the calls to print the result of the Get methods.

Refactor the itkIsolatedWatershedImageFilterTest.cxx so that the close
thresholds condition is exercised by specifying the corresponding test
input parameters through the CMakeLists.txt. Use the MD5 hash for the
newly added test baseline comparison.

Move the tests' Baselines from the general ITK
Testing/Data/Baseline/Algorithms and Testing/Data/Baseline/BasicFilters to
the module's Baseline folder. This allows to reduce the data downloaded
depending on the modules activated.

Improve the style of the tests: use of white spaces to be consistent with
the ITK style, indentation, naming of variables, declaration of the image
dimension as a const, etc.

Change-Id: I23da6fb6aef94cfee5d11dd67f7a71a99bc36589