BUG: Fix frequency iterator to ignore Origin(spatial)
Also adds some frequency metadata

FreqOrigin = 0.0
SamplingFreq =  1 / Spacing
FreqSpacing = SamplingFreq / ImageSize

TODO: Change the name to reflect that this is only for FFT with standard
layout.
Other layouts can be:
 - HalfHermitian
FreqSpacing = SamplingFreq/(2*ImageSize)
 - ShiftedFFT
FreqOrigin = minimun negative frequency?
 - input image taken directly in the frequency domain.
In the later case, the image comes from a photo detector in a scattering
experiment for example, recording the magnitude of the FFT.
GetFrequency in this case would be just calling
TransformIndexToPhysicalPoint()
FreqOrigin = Origin
SamplingFreq = NAN. It doesn't affect the result. Would it be the laser frequency??
FreqSpacing = Spacing