STYLE: Improve the itkRobustAutomaticThreshold classes style.
Use initialization lists.

Make PrintSelf be the last method of the implementation file.

Use the itk::NumericTraits::PrintSelf and the itkPrintSelfObjectMacro to
improve how the ivars and objects that can be null pointers are printed.

Throw exception if the input or gradient images are not set when calling
the itk::RobustAutomaticThresholdCalculator::Compute() function.

Remove the IJ reference from the implementation file; it already exists in
the class documentation.

Remove commented, unnecessary std::cout message.

Some minor white space improvements for the sake of coherence.

Change-Id: I5bc9ab9b0bb5049c0d722e5bfd28d4c2f444fe52