ENH: Add Python convenience function to read images
ITK Python provides a convenience function to write images:
write(imageOrFilter, fileName, compression=False)

A similar function is created to read images:
read(fileName)

This allows to keep Python script shorter, simpler, and have
a matching 'read' function for the 'write' function.

The new 'read()' function also accept an optional argument to specify the
PixelType of the input image. Dimension of the image is found automatically.

Change-Id: I8e6788986bdd6b279e7ddc5d3d17efed490674ba